Added checkbox drop down menu in the editor which controls render flags. Implemented generic Shader which replaces most of the other specialized shaders.

+++ Runtime/Graphics/Material.h
@@ -78,8 +78,8 @@ namespace Directus
 
 		//= SHADER ====================================================================
 		void AcquireShader();
-		std::weak_ptr<ShaderVariation> FindMatchingShader(unsigned int shaderFlags);
-		std::weak_ptr<ShaderVariation> CreateShaderBasedOnMaterial(unsigned int shaderFlags);
+		std::weak_ptr<ShaderVariation> FindMatchingShader(unsigned long shaderFlags);
+		std::weak_ptr<ShaderVariation> CreateShaderBasedOnMaterial(unsigned long shaderFlags);
 		std::weak_ptr<ShaderVariation>& GetShader() { return m_shader; }
 		bool HasShader() { return GetShader().expired() ? false : true; }
 		void** GetShaderResource(TextureType type);
